Add static and client-server combined components
